Factor out pure Node.js module from Grunt task
* Ensure the "src/" directory is included in the package for npm.
* Expose the bananaChecker() function via require('banana-checker'),
  by setting it as script as "main" in package.json.

Fixes #13.
